+++ openeo_driver/jobregistry.py
@@ -17,8 +17,7 @@ from typing import Any, Dict, List, Optional, Sequence, Union
 
 import requests
 import reretry
-from deprecated.classic import deprecated
-from openeo.util import TimingLogger, repr_truncate, rfc3339, url_join
+from openeo.util import TimingLogger, dict_no_none, repr_truncate, rfc3339, url_join
 
 import openeo_driver._version
 from openeo_driver.backend import BatchJobMetadata, JobListing
@@ -142,7 +141,12 @@ class JobRegistryInterface:
         user_id: Optional[str] = None,
         costs: Optional[float],
         usage: dict,
-        results_metadata: Dict[str, Any],
+        results_metadata: Dict[str, Any] = None,
+    ) -> None:
+        raise NotImplementedError
+
+    def set_results_metadata_uri(
+        self, job_id: str, *, user_id: Optional[str] = None, results_metadata_uri: str
     ) -> None:
         raise NotImplementedError
 
@@ -610,6 +614,11 @@ class ElasticJobRegistry(JobRegistryInterface):
     def set_application_id(self, job_id: str, *, user_id: Optional[str] = None, application_id: str) -> None:
         self._update(job_id=job_id, data={"application_id": application_id})
 
+    def set_results_metadata_uri(
+        self, job_id: str, *, user_id: Optional[str] = None, results_metadata_uri: str
+    ) -> None:
+        self._update(job_id=job_id, data={"results_metadata_uri": results_metadata_uri})
+
     def _search(self, query: dict, fields: Optional[List[str]] = None) -> List[JobDict]:
         # TODO: sorting, pagination?
         fields = set(fields or [])
@@ -747,15 +756,17 @@ class ElasticJobRegistry(JobRegistryInterface):
         user_id: Optional[str] = None,
         costs: Optional[float],
         usage: dict,
-        results_metadata: Dict[str, Any],
+        results_metadata: Dict[str, Any] = None,
     ) -> None:
         self._update(
             job_id=job_id,
-            data={
-                "costs": costs,
-                "usage": usage,
-                "results_metadata": results_metadata,
-            },
+            data=dict_no_none(
+                {
+                    "costs": costs,
+                    "usage": usage,
+                    "results_metadata": results_metadata,
+                }
+            ),
         )